refactor(ecs): move query to system dir
